% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayPlotGadget.R
\name{arrayPlot}
\alias{arrayPlot}
\title{Array plot function to work on ST data}
\description{
Opens up a gadet, put your gene name to display the counts along the tissue.
Your count matrix needs to be in the form samples -> columns, genes -> rows.
The columns names needs to be in the form of [x]x[y], where [x] and [y] are coordinates on the array.
}
\examples{
arrayPlot(counts)
}
\keyword{arrayplot}
